+++ examples/demo/global.R
@@ -24,3 +24,9 @@ credentials <- data.frame(
 #   passphrase = "supersecret"
 # )
 
+set_labels(
+  language = "en",
+  "Please authenticate" = "You have to login",
+  "Username:" = "What's your name:",
+  "Password:" = "Enter your password:"
+)